feat: modify migration with companies table and modifying migration column due to error by rollbacking migration

+++ database/migrations/2023_05_04_172510_create_companies_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->foreignId('industry_id')->nullable();
             $table->string('name')->unique();
             $table->string('slug')->unique();
-            $table->unsignedInteger('vat_number')->unique();
+            $table->unsignedBigInteger('vat_number')->unique();
             $table->unsignedInteger('status')->default(0);
             $table->text('description')->nullable();
             $table->string('employees')->nullable();
