+++ database/migrations/2023_05_05_031234_modify_vat_number_in_companies_table.php
@@ -22,7 +22,7 @@ return new class extends Migration
     public function down(): void
     {
         Schema::table('companies', function (Blueprint $table) {
-            $table->unsignedInteger('vat_number')->change();
+            $table->unsignedBigInteger('vat_number')->change();
         });
     }
 };
